requirements.txt
numpy
pillow
pyfiglet